Check state before automatic on/off
The state check fixes an issue where the log was being spam during each
cycle during a 1 second period because the auto on/of function was
repeatedly trying to set the lights' state.

Also added the ability to configure the minute at which the lights turn
on or off to the config file.